Feat: Company info
- caching, showing just basic info - should be improved

+++ app/src/main/java/cz/stepanzalis/spacexlifts/io/response/CompanyInfoResponse.kt
@@ -32,7 +32,7 @@ data class CompanyInfoResponse(
     @Json(name = "cto_propulsion")
     val ctoPropulsion: String,
     @Json(name = "valuation")
-    val valuation: Int,
+    val valuation: Long, // when valuation is too big that int is not enough 😱
     @Json(name = "summary")
     val summary: String,
     @Json(name = "id")
